+++ beacon_chain/gossip_processing/block_processor.nim
@@ -941,7 +941,7 @@ proc storeBackfillPayload(
   self.storeSidecars(sidecarsOpt)
   ok()
 
-proc addPayload*(
+proc addPayload(
     self: ref BlockProcessor,
     signedBlock: gloas.SignedBeaconBlock,
     signedEnvelope: gloas.SignedExecutionPayloadEnvelope,
@@ -993,9 +993,14 @@ proc addPayload*(
   # TODO To be removed - Temporary call without import.
   blockchain_dag.updateHeadExecutionPayload(dag, blck, signedEnvelope)
 
+  debug "Envelope processed",
+    head = shortLog(dag.head),
+    blck = shortLog(blck),
+    slot = signedBlock.message.slot
+
   # Store sidecars into db.
   self[].storeSidecars(sidecarsOpt)
-  self.envelopeQuarantine[].delOrphan(signedBlock)
+  self.envelopeQuarantine[].remove(signedBlock.root)
 
   ok()