+++ beacon_chain/nimbus_beacon_node.nim
@@ -23,7 +23,6 @@ import
   ./networking/[topic_params, network_metadata_downloads],
   ./rpc/[rest_api, state_ttl_cache],
   ./el/el_getblobs_service,
-  ./spec/datatypes/[altair, bellatrix, phase0],
   ./spec/[
     engine_authentication, weak_subjectivity, peerdas_helpers],
   ./sync/[sync_protocol, light_client_protocol, sync_overseer, validator_custody],
@@ -493,6 +492,8 @@ proc initFullNode(
       else:
         data
     node.eventBus.reorgQueue.emit(eventData)
+  proc onEnvelopeAdded(data: ExecutionPayloadInfoObject) =
+    node.eventBus.execPayloadAvlQueue.emit(data)
   proc makeOnFinalizationCb(
       # This `nimcall` functions helps for keeping track of what
       # needs to be captured by the onFinalization closure.
@@ -552,7 +553,7 @@ proc initFullNode(
   let
     quarantine = newClone(
       Quarantine.init(dag.cfg))
-    envelopeQuarantine = newClone(EnvelopeQuarantine.init())
+    envelopeQuarantine = newClone(EnvelopeQuarantine.init(onEnvelopeAdded))
     attestationPool = newClone(AttestationPool.init(
       dag, quarantine, getBeaconTime(), onSingleAttestationReceived))
     syncCommitteeMsgPool = newClone(
@@ -607,18 +608,17 @@ proc initFullNode(
       batchVerifier, consensusManager, node.validatorMonitor,
       blobQuarantine, dataColumnQuarantine, gloasColumnQuarantine,
       envelopeQuarantine, getBeaconTime, config.invalidBlockRoots)
-    blockVerifier = proc(
-        signedBlock: ForkedSignedBeaconBlock,
-        signedEnvelope: Opt[ref SignedExecutionPayloadEnvelope],
-        blobs: Opt[BlobSidecars], maybeFinalized: bool):
-        Future[Result[void, VerifierError]] {.async: (raises: [CancelledError]).} =
+    blockVerifier = proc(signedBlock: ForkedSignedBeaconBlock,
+                         blobs: Opt[BlobSidecars], maybeFinalized: bool):
+        Future[Result[void, VerifierError]] {.async: (raises: [CancelledError], raw: true).} =
       withBlck(signedBlock):
-        when consensusFork >= ConsensusFork.Gloas:
-          # Disable sidecars processing at block time.
-          const sidecarsOpt = noSidecars
-        elif consensusFork == ConsensusFork.Fulu:
+        when consensusFork in ConsensusFork.Fulu .. ConsensusFork.Gloas:
           # TODO document why there are no columns here
-          let sidecarsOpt = Opt.none(fulu.DataColumnSidecars)
+          when consensusFork == ConsensusFork.Gloas:
+            # Disable sidecars processing at block time.
+            const sidecarsOpt = noSidecars
+          else:
+            let sidecarsOpt = Opt.none(fulu.DataColumnSidecars)
         elif consensusFork in ConsensusFork.Deneb .. ConsensusFork.Electra:
           template sidecarsOpt: untyped = blobs
         elif consensusFork in ConsensusFork.Phase0 .. ConsensusFork.Capella:
@@ -626,36 +626,14 @@ proc initFullNode(
         else:
           {.error: "Unkown fork: " & $consensusFork.}
 
-        let bres = await blockProcessor.addBlock(
+        blockProcessor.addBlock(
           MsgSource.gossip, forkyBlck, sidecarsOpt, maybeFinalized)
 
-        when consensusFork >= ConsensusFork.Gloas:
-          template bid(): auto =
-            forkyBlck.message.body.signed_execution_payload_bid
-          if bres.isErr():
-            bres
-          elif signedEnvelope.isNone():
-            err(VerifierError.Invalid)
-          else:
-            let columnsOpt =
-              if len(bid.message.blob_kzg_commitments) > 0:
-                gloasColumnQuarantine[].popSidecars(forkyBlck.root)
-              else:
-                Opt.some(default(gloas.DataColumnSidecars))
-
-            debugGloasComment("columns may not be guaranteed")
-            await blockProcessor.addPayload(
-              forkyBlck, signedEnvelope.get()[], columnsOpt)
-        else:
-          bres
-
-    untrustedBlockVerifier = proc(
-        signedBlock: ForkedSignedBeaconBlock,
-        signedEnvelope: Opt[ref SignedExecutionPayloadEnvelope],
-        blobs: Opt[BlobSidecars], maybeFinalized: bool):
-        Future[Result[void, VerifierError]] {.async: (raises: [CancelledError], raw: true).} =
-      debugGloasComment("")
-      clist.untrustedBackfillVerifier(signedBlock, blobs, maybeFinalized)
+    untrustedBlockVerifier =
+      proc(signedBlock: ForkedSignedBeaconBlock, blobs: Opt[BlobSidecars],
+           maybeFinalized: bool): Future[Result[void, VerifierError]] {.
+        async: (raises: [CancelledError], raw: true).} =
+        clist.untrustedBackfillVerifier(signedBlock, blobs, maybeFinalized)
     rmanBlockVerifier = proc(signedBlock: ForkedSignedBeaconBlock,
                              maybeFinalized: bool):
         Future[Result[void, VerifierError]] {.async: (raises: [CancelledError]).} =
@@ -730,7 +708,6 @@ proc initFullNode(
       node.network.peerPool,
       dag.cfg.DENEB_FORK_EPOCH,
       dag.cfg.FULU_FORK_EPOCH,
-      dag.cfg.GLOAS_FORK_EPOCH,
       dag.cfg.MIN_EPOCHS_FOR_BLOB_SIDECARS_REQUESTS,
       dag.cfg.MAX_BLOBS_PER_BLOCK_ELECTRA,
       SyncQueueKind.Forward, getLocalHeadSlot,
@@ -743,7 +720,6 @@ proc initFullNode(
       node.network.peerPool,
       dag.cfg.DENEB_FORK_EPOCH,
       dag.cfg.FULU_FORK_EPOCH,
-      dag.cfg.GLOAS_FORK_EPOCH,
       dag.cfg.MIN_EPOCHS_FOR_BLOB_SIDECARS_REQUESTS,
       dag.cfg.MAX_BLOBS_PER_BLOCK_ELECTRA,
       SyncQueueKind.Backward, getLocalHeadSlot,
@@ -762,7 +738,6 @@ proc initFullNode(
       node.network.peerPool,
       dag.cfg.DENEB_FORK_EPOCH,
       dag.cfg.FULU_FORK_EPOCH,
-      dag.cfg.GLOAS_FORK_EPOCH,
       dag.cfg.MIN_EPOCHS_FOR_BLOB_SIDECARS_REQUESTS,
       dag.cfg.MAX_BLOBS_PER_BLOCK_ELECTRA,
       SyncQueueKind.Backward, getLocalHeadSlot,
@@ -2343,8 +2318,7 @@ proc installMessageValidators(node: BeaconNode) =
               src: PeerId
             ): ValidationResult =
               toValidationResult(
-                node.processor[].processExecutionPayloadBid(
-                  MsgSource.gossip, signedBid)))
+                node.processor[].processExecutionPayloadBid(signedBid)))
 
         # execution_payload
         # https://github.com/ethereum/consensus-specs/blob/v1.7.0-alpha-1/specs/gloas/p2p-interface.md#execution_payload
@@ -2370,8 +2344,8 @@ proc installMessageValidators(node: BeaconNode) =
                  async: (raises: [CancelledError]).} =
               return toValidationResult(
                 await node.processor.processPayloadAttestationMessage(
-                  MsgSource.gossip, payloadAttestationMessage,
-                  checkSignature = true, checkValidator = false)))
+                  payloadAttestationMessage, checkSignature = true,
+                  checkValidator = false)))
 
         # beacon_attestation_{subnet_id}
         # https://github.com/ethereum/consensus-specs/blob/v1.4.0-beta.5/specs/phase0/p2p-interface.md#beacon_attestation_subnet_id