+++ beacon_chain/sync/sync_manager.nim
@@ -56,7 +56,6 @@ type
     pool: PeerPool[A, B]
     DENEB_FORK_EPOCH: Epoch
     FULU_FORK_EPOCH: Epoch
-    GLOAS_FORK_EPOCH: Epoch
     MIN_EPOCHS_FOR_BLOB_SIDECARS_REQUESTS: uint64
     MAX_BLOBS_PER_BLOCK_ELECTRA: uint64
     responseTimeout: chronos.Duration
@@ -94,14 +93,11 @@ type
 
   BeaconBlocksRes =
     NetRes[List[ref ForkedSignedBeaconBlock, Limit MAX_REQUEST_BLOCKS]]
-  EnvelopesRes =
-    NetRes[List[ref SignedExecutionPayloadEnvelope, Limit MAX_REQUEST_BLOCKS]]
   BlobSidecarsRes =
     NetRes[List[ref BlobSidecar, Limit(MAX_SUPPORTED_REQUEST_BLOB_SIDECARS)]]
 
   SyncBlockData* = object
     blocks*: seq[ref ForkedSignedBeaconBlock]
-    envelopes*: Opt[seq[ref SignedExecutionPayloadEnvelope]]
     blobs*: Opt[seq[BlobSidecars]]
 
   SyncBlockDataRes* = Result[SyncBlockData, string]
@@ -149,7 +145,6 @@ proc newSyncManager*[A, B](
     pool: PeerPool[A, B],
     denebEpoch: Epoch,
     fuluEpoch: Epoch,
-    gloasEpoch: Epoch,
     minEpochsForBlobSidecarsRequests: uint64,
     maxBlobsPerBlockElectra: uint64,
     direction: SyncQueueKind,
@@ -181,7 +176,6 @@ proc newSyncManager*[A, B](
     pool: pool,
     DENEB_FORK_EPOCH: denebEpoch,
     FULU_FORK_EPOCH: fuluEpoch,
-    GLOAS_FORK_EPOCH: gloasEpoch,
     MIN_EPOCHS_FOR_BLOB_SIDECARS_REQUESTS: minEpochsForBlobSidecarsRequests,
     MAX_BLOBS_PER_BLOCK_ELECTRA: maxBlobsPerBlockElectra,
     getLocalHeadSlot: getLocalHeadSlotCb,
@@ -220,20 +214,6 @@ proc getBlocks[A, B](man: SyncManager[A, B], peer: A,
         topics = "syncman"
 
   beaconBlocksByRange_v2(peer, req.data.slot, req.data.count, 1'u64)
-
-proc getEnvelopes[A, B](man: SyncManager[A, B], peer: A,
-                        req: SyncRequest[A]): Future[EnvelopesRes] {.
-                        async: (raises: [CancelledError], raw: true).} =
-  mixin getScore, `==`
-  doAssert(not(req.isEmpty()), "Request must not be empty!")
-  debug "Requesting envelopes from peer",
-        request = req,
-        peer_score = req.item.getScore(),
-        peer_speed = req.item.netKbps(),
-        sync_ident = man.ident,
-        topics = "syncman"
-
-  executionPayloadEnvelopesByRange(peer, req.data.slot, req.data.count)
 
 proc shouldGetBlobs[A, B](man: SyncManager[A, B], s: Slot): bool =
   let
@@ -365,43 +345,16 @@ proc getSyncBlockData*[T](
     peer.updateScore(PeerScoreBadResponse)
     return err("The received block is not in the requested range")
 
-  let (shouldGetEnvelope, shouldGetBlob, blobsCount) =
+  let (shouldGetBlob, blobsCount) =
     withBlck(blocksRange[0][]):
-      when consensusFork == ConsensusFork.Gloas:
-        (true, false, 0)
-      elif consensusFork in [ConsensusFork.Deneb, ConsensusFork.Electra]:
+      when consensusFork in [ConsensusFork.Deneb, ConsensusFork.Electra]:
         let res = len(forkyBlck.message.body.blob_kzg_commitments)
         if res > 0:
-          (false, true, res)
+          (true, res)
         else:
-          (false, false, 0)
+          (false, 0)
       else:
-        (false, false, 0)
-
-  let envelopeRange =
-    if shouldGetEnvelope:
-      let envelopeRange =
-        block:
-          let res = await executionPayloadEnvelopesByRange(peer, slot, 1'u64)
-          if res.isErr():
-            peer.updateScore(PeerScoreNoValues)
-            return err(
-              "Failed to receive envelopes on request, reason: " & $res.error)
-          res.get().asSeq()
-
-      if len(envelopeRange) == 0:
-        peer.updateScore(PeerScoreNoValues)
-        return err("An empty range of envelopes was returned by peer")
-
-      if len(envelopeRange) != 1:
-        peer.updateScore(PeerScoreBadResponse)
-        return err(
-          "Incorrect number of envelopes was returned by peer, " &
-          $len(envelopeRange))
-
-      Opt.some(envelopeRange)
-    else:
-      Opt.none(seq[ref SignedExecutionPayloadEnvelope])
+        (false, 0)
 
   let blobsRange =
     if shouldGetBlob:
@@ -449,9 +402,7 @@ proc getSyncBlockData*[T](
     else:
       Opt.none(seq[BlobSidecars])
 
-  ok(SyncBlockData(
-    blocks: blocksRange, envelopes: envelopeRange, blobs: blobsRange,
-  ))
+  ok(SyncBlockData(blocks: blocksRange, blobs: blobsRange))
 
 proc getSyncBlockData[A, B](
     man: SyncManager[A, B],
@@ -528,22 +479,8 @@ proc getSyncBlockData[A, B](
         Opt.some(groupedBlobs)
       else:
         Opt.none(seq[BlobSidecars])
-    shouldGetEnvelope = sr.data.slot.epoch() >= man.GLOAS_FORK_EPOCH
-    envelopes =
-      if shouldGetEnvelope:
-        let envelopes = (await man.getEnvelopes(peer, sr)).valueOr:
-          peer.updateScore(PeerScoreNoValues)
-          return err(
-            "Failed to receive envelopes on request, reason: " & $error)
 
-        debugGloasComment("verify response")
-        Opt.some(envelopes.asSeq())
-      else:
-        Opt.none(seq[ref SignedExecutionPayloadEnvelope])
-
-  ok(SyncBlockData(
-    blocks: blocks.asSeq(), envelopes: envelopes, blobs: blobs,
-  ))
+  ok(SyncBlockData(blocks: blocks.asSeq(), blobs: blobs))
 
 proc getOrUpdatePeerStatus[A, B](
     man: SyncManager[A, B], index: int, peer: A
@@ -768,8 +705,8 @@ proc syncStep[A, B](
         # TODO descore peers that lie
         maybeFinalized = lastSlot < peerFinalized
 
-      jobs.add(man.queue.push(request, data.blocks, data.envelopes, data.blobs,
-                              maybeFinalized, processCallback))
+      jobs.add(man.queue.push(request, data.blocks, data.blobs, maybeFinalized,
+                              processCallback))
 
     if len(jobs) > 0:
       await allFutures(jobs)