+++ beacon_chain/sync/sync_queue.nim
@@ -22,11 +22,8 @@ type
   GetSlotCallback* = proc(): Slot {.gcsafe, raises: [].}
   GetBoolCallback* = proc(): bool {.gcsafe, raises: [].}
   ProcessingCallback* = proc() {.gcsafe, raises: [].}
-  BlockVerifier* =
-    proc(
-      signedBlock: ForkedSignedBeaconBlock,
-      signedEnvelope: Opt[ref SignedExecutionPayloadEnvelope],
-      blobs: Opt[BlobSidecars], maybeFinalized: bool):
+  BlockVerifier* =  proc(signedBlock: ForkedSignedBeaconBlock,
+                         blobs: Opt[BlobSidecars], maybeFinalized: bool):
       Future[Result[void, VerifierError]] {.async: (raises: [CancelledError]).}
   ForkAtEpochCallback* =
     proc(epoch: Epoch): ConsensusFork {.gcsafe, raises: [].}
@@ -744,29 +741,18 @@ func getOpt(blobs: Opt[seq[BlobSidecars]], i: int): Opt[BlobSidecars] =
   else:
     Opt.none(BlobSidecars)
 
-func getOpt(
-    envelopes: Opt[seq[ref gloas.SignedExecutionPayloadEnvelope]], i: int
-): Opt[ref gloas.SignedExecutionPayloadEnvelope] =
-  if envelopes.isSome:
-    Opt.some(envelopes.get()[i])
-  else:
-    Opt.none(ref gloas.SignedExecutionPayloadEnvelope)
-
 iterator blocks(
     kind: SyncQueueKind,
     blcks: seq[ref ForkedSignedBeaconBlock],
-    envelopes: Opt[seq[ref gloas.SignedExecutionPayloadEnvelope]],
     blobs: Opt[seq[BlobSidecars]]
-): (ref ForkedSignedBeaconBlock,
-    Opt[ref gloas.SignedExecutionPayloadEnvelope],
-    Opt[BlobSidecars]) =
+): (ref ForkedSignedBeaconBlock, Opt[BlobSidecars]) =
   case kind
   of SyncQueueKind.Forward:
     for i in countup(0, len(blcks) - 1):
-      yield (blcks[i], envelopes.getOpt(i), blobs.getOpt(i))
+      yield (blcks[i], blobs.getOpt(i))
   of SyncQueueKind.Backward:
     for i in countdown(len(blcks) - 1, 0):
-      yield (blcks[i], envelopes.getOpt(i), blobs.getOpt(i))
+      yield (blcks[i], blobs.getOpt(i))
 
 proc push*[T](sq: SyncQueue[T], requests: openArray[SyncRequest[T]]) =
   ## Push multiple failed requests back to queue.
@@ -784,7 +770,6 @@ proc process[T](
     sq: SyncQueue[T],
     sr: SyncRequest[T],
     blcks: seq[ref ForkedSignedBeaconBlock],
-    envelopes: Opt[seq[ref gloas.SignedExecutionPayloadEnvelope]],
     blobs: Opt[seq[BlobSidecars]],
     maybeFinalized: bool
 ): Future[SyncProcessingResult] {.
@@ -797,8 +782,8 @@ proc process[T](
   if len(blcks) == 0:
     return SyncProcessingResult.init(SyncProcessError.Empty)
 
-  for blk, evl, blb in blocks(sq.kind, blcks, envelopes, blobs):
-    let res = await sq.blockVerifier(blk[], evl, blb, maybeFinalized)
+  for blk, blb in blocks(sq.kind, blcks, blobs):
+    let res = await sq.blockVerifier(blk[], blb, maybeFinalized)
     if res.isOk():
       slot = Opt.some(SyncBlock.init(blk[].slot, blk[].root))
     else:
@@ -844,7 +829,6 @@ proc push*[T](
     sq: SyncQueue[T],
     sr: SyncRequest[T],
     data: seq[ref ForkedSignedBeaconBlock],
-    envelopes: Opt[seq[ref gloas.SignedExecutionPayloadEnvelope]],
     blobs: Opt[seq[BlobSidecars]],
     maybeFinalized: bool = false,
     processingCb: ProcessingCallback = nil
@@ -899,7 +883,7 @@ proc push*[T](
     if not(isNil(processingCb)):
       processingCb()
 
-    let pres = await sq.process(sr, data, envelopes, blobs, maybeFinalized)
+    let pres = await sq.process(sr, data, blobs, maybeFinalized)
 
     # We need to update position, because while we waiting for `process()` to
     # complete - clearAndWakeup() could be invoked which could clean whole the