+++ flake.nix
@@ -3,10 +3,6 @@
 
   inputs = {
     nixpkgs.url = "github:NixOS/nixpkgs?rev=2a777ace4b722f2714cc06d596f2476ee628c04a";
-    nimbusBuildSystem = {
-      url = "git+file:./vendor/nimbus-build-system?submodules=1";
-      inputs.nixpkgs.follows = "nixpkgs";
-    };
     self = {
       # WARNING: Does not work with 'github:' schema URLs.
       # https://github.com/NixOS/nix/issues/14982
@@ -16,7 +12,7 @@
     };
   };
 
-  outputs = { self, nixpkgs, nimbusBuildSystem }:
+  outputs = { self, nixpkgs }:
     assert (builtins.compareVersions builtins.nixVersion "2.27") <= 0
       -> throw "Nix 2.27 or newer needed for proper submodules support!";
 
@@ -36,8 +32,7 @@
         buildTarget = pkgsFor.${system}.callPackage ./nix/default.nix {
           inherit stableSystems; src = self;
         };
-        nim = nimbusBuildSystem.packages.${system}.nim;
-        build = targets: buildTarget.override { inherit targets nim; };
+        build = targets: buildTarget.override { inherit targets; };
       in rec {
         beacon_node      = build ["nimbus_beacon_node"];
         signing_node     = build ["nimbus_signing_node"];